# docs/openapi-draft.yaml
openapi: 3.1.0
info:
  title: Arbibot 2 — sync API draft (Phase 0)
  version: 0.1.0
  description: >
    Черновик REST контрактов первой очереди (P0-0.1-OAPI).
    Реализации — в сервисах; типы DTO дублируют схемы в коде.

paths:
  /evaluate-risk:
    post:
      operationId: evaluateRisk
      summary: Оценка риска по возможности/плану
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluateRiskRequest'
      responses:
        '200':
          description: Решение создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluateRiskResponse'
        '400':
          description: Ошибка валидации

  /risk-decisions/{id}:
    get:
      operationId: getRiskDecision
      summary: Получить решение по id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskDecision'
        '404':
          description: Не найдено

  /opportunities:
    get:
      operationId: listOpportunities
      summary: Список возможностей (read)
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ArbitrageOpportunity'
    post:
      operationId: createOpportunity
      summary: Зарегистрировать возможность (ingest/strategy)
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOpportunityRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArbitrageOpportunity'

  /opportunities/{id}:
    get:
      operationId: getOpportunity
      summary: Получить возможность по id (read, с payload)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArbitrageOpportunityDetail'
        '404':
          description: Не найдено

  /opportunities/{id}/enrich:
    post:
      operationId: enrichOpportunity
      summary: Перевод detected → enriched, опционально patch payload
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrichOpportunityRequest'
      responses:
        '200':
          description: Обновлённая возможность
        '400':
          description: Неверное состояние / валидация
        '404':
          description: Не найдено
        '409':
          description: Конфликт состояния

  /opportunities/{id}/request-risk-evaluation:
    post:
      operationId: requestRiskEvaluation
      summary: Вызов risk-service (evaluate-risk), переход в risk_checked
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestRiskEvaluationRequest'
      responses:
        '200':
          description: Результат оценки / idempotent replay
        '400':
          description: Валидация
        '404':
          description: Возможность не найдена
        '409':
          description: Конфликт состояния / idempotency
        '503':
          description: Risk-service недоступен

  /opportunities/{id}/paper-enqueue:
    post:
      operationId: enqueuePaperPromotion
      summary: Запись в outbox на создание кандидата promotion в paper-trading-service (доставка relay)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaperEnqueueRequest'
      responses:
        '200':
          description: Строка outbox записана (relay доставит в paper при доступном PAPER_TRADING_SERVICE_URL)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaperEnqueueResponse'
        '404':
          description: Возможность не найдена

  /paper/trades:
    get:
      operationId: listPaperTrades
      summary: Список paper trades (read)
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PaperTrade'
    post:
      operationId: createPaperTrade
      summary: Создать paper trade (опционально idempotencyKey)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaperTradeRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaperTrade'

  /paper/trades/{id}:
    get:
      operationId: getPaperTrade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaperTrade'
        '404':
          description: Не найдено
    patch:
      operationId: patchPaperTrade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchPaperTradeRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaperTrade'
        '404':
          description: Не найдено
        '409':
          description: Версия или недопустимый переход state

  /paper/promotion-candidates:
    get:
      operationId: listPaperPromotionCandidates
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PaperPromotionCandidate'
    post:
      operationId: createPaperPromotionCandidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaperPromotionCandidateRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaperPromotionCandidate'

  /paper/promotion-candidates/{id}:
    patch:
      operationId: patchPaperPromotionCandidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchPaperPromotionCandidateRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaperPromotionCandidate'

  /capital/reservations:
    post:
      operationId: reserveCapital
      summary: Зарезервировать капитал (reservation-first)
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReserveCapitalRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapitalReservation'

  /execution/plans:
    post:
      operationId: createExecutionPlan
      summary: Создать план (planned)
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionPlan'
    get:
      operationId: listExecutionPlans
      summary: Список планов
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ExecutionPlan'

  /execution/plans/{id}/arm:
    post:
      operationId: armPlan
      summary: Arm плана при валидных резервах
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionPlan'

  /market/resolve-instrument:
    post:
      operationId: resolveInstrument
      summary: Разрешить venue symbol → канонический инструмент
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveInstrumentRequest'
      responses:
        '200':
          description: Найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResolvedInstrument'
        '400':
          description: Неверная комбинация полей
        '404':
          description: Площадка или инструмент не найдены

  /market/resolve-route:
    post:
      operationId: resolveRoute
      summary: Разрешить маршрут по ключу или паре инструментов
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveRouteRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResolvedRoute'
        '400':
          description: Неверная комбинация полей
        '404':
          description: Маршрут не найден
        '409':
          description: Несколько маршрутов на одну пару инструментов — укажите routeKey

  /snapshots/ingest:
    post:
      operationId: ingestMarketSnapshot
      summary: Принять нормализованный снимок рынка (single-writer intake)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestMarketSnapshotRequest'
      responses:
        '200':
          description: >
            Снимок обработан. Новая запись outbox SnapshotUpdated (schema v2) только если состояние изменилось.
            Заголовок X-Idempotent-Replayed=true при повторе с тем же idempotencyKey и тем же телом.
          headers:
            X-Idempotent-Replayed:
              schema:
                type: string
                enum: ['true']
              description: Повтор идемпотентного запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestMarketSnapshotResponse'
        '400':
          description: Валидация
        '409':
          description: Конфликт idempotencyKey и тела запроса

  /snapshots:
    get:
      operationId: getMarketSnapshot
      summary: Получить последний снимок по venue + symbol (read)
      parameters:
        - name: venueCode
          in: query
          required: true
          schema:
            type: string
        - name: venueSymbol
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketSnapshotReadResponse'
        '404':
          description: Нет данных

components:
  schemas:
    EvaluateRiskRequest:
      type: object
      required: [correlationId, planReference, notionalUsd, snapshotVersion]
      properties:
        correlationId:
          type: string
        planReference:
          type: string
        notionalUsd:
          type: number
        snapshotVersion:
          type: integer
        riskMode:
          type: string
          enum: [fast, standard, conservative]
          default: standard

    EvaluateRiskResponse:
      type: object
      required: [id, outcome, correlationId]
      properties:
        id:
          type: string
          format: uuid
        outcome:
          type: string
          enum: [approved, rejected, deferred]
        correlationId:
          type: string
        reasons:
          type: array
          items:
            type: string

    PaperEnqueueRequest:
      type: object
      properties:
        instrumentKey:
          type: string
        score:
          type: number
        driftBps:
          type: number
        evidence:
          type: object

    PaperEnqueueResponse:
      type: object
      required: [enqueued, paperServiceConfigured]
      properties:
        enqueued:
          type: boolean
          description: true если запись в outbox создана (relay доставит в paper)
        paperServiceConfigured:
          type: boolean

    CreatePaperTradeRequest:
      type: object
      required: [instrumentKey]
      properties:
        opportunityId:
          type: string
          format: uuid
        instrumentKey:
          type: string
        routeKey:
          type: string
        notional:
          type: string
        summary:
          type: object
        idempotencyKey:
          type: string

    PatchPaperTradeRequest:
      type: object
      required: [expectedVersion]
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        state:
          type: string
          enum: [draft, active, settled, canceled]
        notional:
          type: string

    PaperTrade:
      type: object
      required: [id, instrumentKey, state, notional, entityVersion, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        opportunityId:
          type: string
          format: uuid
          nullable: true
        instrumentKey:
          type: string
        routeKey:
          type: string
          nullable: true
        state:
          type: string
          enum: [draft, active, settled, canceled]
        notional:
          type: string
        summary:
          type: object
        entityVersion:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreatePaperPromotionCandidateRequest:
      type: object
      required: [instrumentKey]
      properties:
        instrumentKey:
          type: string
        opportunityId:
          type: string
          format: uuid
        source:
          type: string
        score:
          type: number
        driftBps:
          type: number
        evidence:
          type: object
        enqueueIdempotencyKey:
          type: string
          description: Стабильный ключ для идемпотентного create (ретраи relay / повторные POST)

    PatchPaperPromotionCandidateRequest:
      type: object
      required: [expectedVersion, status]
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [queued, under_review, promoted, rejected, expired]

    PaperPromotionCandidate:
      type: object
      required: [id, instrumentKey, source, status, evidence, entityVersion, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        instrumentKey:
          type: string
        opportunityId:
          type: string
          format: uuid
          nullable: true
        source:
          type: string
        status:
          type: string
        score:
          type: string
          nullable: true
        driftBps:
          type: string
          nullable: true
        evidence:
          type: object
        enqueueIdempotencyKey:
          type: string
          nullable: true
        entityVersion:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    RiskDecision:
      allOf:
        - $ref: '#/components/schemas/EvaluateRiskResponse'
        - type: object
          properties:
            planReference:
              type: string
            snapshotVersion:
              type: integer
            entityVersion:
              type: integer
            createdAtIso:
              type: string
              format: date-time

    ArbitrageOpportunity:
      type: object
      properties:
        id:
          type: string
          format: uuid
        state:
          type: string
        correlationId:
          type: string
        entityVersion:
          type: integer

    ArbitrageOpportunityDetail:
      allOf:
        - $ref: '#/components/schemas/ArbitrageOpportunity'
        - type: object
          properties:
            riskDecisionId:
              type: string
              format: uuid
            payload:
              type: object
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time

    EnrichOpportunityRequest:
      type: object
      properties:
        payloadPatch:
          type: object

    RequestRiskEvaluationRequest:
      type: object
      required: [notionalUsd, snapshotVersion]
      properties:
        correlationId:
          type: string
          format: uuid
        idempotencyKey:
          type: string
          format: uuid
        notionalUsd:
          type: number
        snapshotVersion:
          type: integer
          minimum: 1
        riskMode:
          type: string
          enum: [fast, standard, conservative]
        riskWindowReservationId:
          type: string
          format: uuid

    CreateOpportunityRequest:
      type: object
      required: [correlationId]
      properties:
        correlationId:
          type: string
        payload:
          type: object

    ReserveCapitalRequest:
      type: object
      required: [correlationId, amountUsd]
      properties:
        correlationId:
          type: string
        planId:
          type: string
          format: uuid
        amountUsd:
          type: number
        ttlSeconds:
          type: integer
          default: 300

    CapitalReservation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        state:
          type: string
        correlationId:
          type: string
        expiresAt:
          type: string
          format: date-time

    ExecutionPlan:
      type: object
      properties:
        id:
          type: string
          format: uuid
        state:
          type: string
        correlationId:
          type: string
        capitalReservationId:
          type: string
          format: uuid
        riskDecisionId:
          type: string
          format: uuid
        entityVersion:
          type: integer

    ResolveInstrumentRequest:
      type: object
      description: Ровно один режим — по canonicalKey или пара venueCode+venueSymbol
      properties:
        canonicalKey:
          type: string
        venueCode:
          type: string
        venueSymbol:
          type: string

    ResolvedInstrument:
      type: object
      required:
        [id, venueCode, venueSymbol, canonicalKey, baseAsset, quoteAsset, entityVersion, createdAtIso, updatedAtIso]
      properties:
        id:
          type: string
          format: uuid
        venueCode:
          type: string
        venueSymbol:
          type: string
        canonicalKey:
          type: string
        baseAsset:
          type: string
        quoteAsset:
          type: string
        attributes:
          type: object
        entityVersion:
          type: integer
        createdAtIso:
          type: string
          format: date-time
        updatedAtIso:
          type: string
          format: date-time

    ResolveRouteRequest:
      type: object
      properties:
        routeKey:
          type: string
        sourceInstrumentId:
          type: string
          format: uuid
        targetInstrumentId:
          type: string
          format: uuid

    ResolvedRoute:
      type: object
      required:
        [id, routeKey, sourceInstrumentId, targetInstrumentId, hops, entityVersion, createdAtIso, updatedAtIso]
      properties:
        id:
          type: string
          format: uuid
        routeKey:
          type: string
        sourceInstrumentId:
          type: string
          format: uuid
        targetInstrumentId:
          type: string
          format: uuid
        hops:
          type: array
          items: {}
        entityVersion:
          type: integer
        createdAtIso:
          type: string
          format: date-time
        updatedAtIso:
          type: string
          format: date-time

    IngestMarketSnapshotRequest:
      type: object
      required: [venueCode, venueSymbol, observedAt]
      properties:
        idempotencyKey:
          type: string
          format: uuid
          description: Опционально; повтор с тем же ключом и тем же телом — тот же результат без второй записи outbox
        correlationId:
          type: string
          format: uuid
        venueCode:
          type: string
        venueSymbol:
          type: string
        canonicalInstrumentId:
          type: string
          format: uuid
        bid:
          type: number
        ask:
          type: number
        last:
          type: number
        payload:
          type: object
        observedAt:
          type: string
          format: date-time
        staleAfterSeconds:
          type: integer
          minimum: 1

    IngestMarketSnapshotResponse:
      type: object
      required: [snapshotId, entityVersion, idempotentReplay, unchanged]
      properties:
        snapshotId:
          type: string
          format: uuid
        outboxMessageId:
          type: string
          format: uuid
          nullable: true
          description: Новый messageId outbox; null если снимок не менялся (stale observedAt / дубликат содержимого)
        entityVersion:
          type: integer
        idempotentReplay:
          type: boolean
        unchanged:
          type: boolean
          description: true если запрос не изменил состояние (устаревший observedAt или идентичное содержимое)

    MarketSnapshotReadResponse:
      type: object
      required: [snapshot, freshness]
      properties:
        snapshot:
          type: object
          properties:
            id:
              type: string
              format: uuid
            venueCode:
              type: string
            venueSymbol:
              type: string
            canonicalInstrumentId:
              type: string
              format: uuid
              nullable: true
            bid:
              type: number
              nullable: true
            ask:
              type: number
              nullable: true
            last:
              type: number
              nullable: true
            payload:
              type: object
            observedAt:
              type: string
              format: date-time
            receivedAt:
              type: string
              format: date-time
            staleAfterSeconds:
              type: integer
              nullable: true
            entityVersion:
              type: integer
        freshness:
          type: object
          required: [observedAt, receivedAt, isStale]
          properties:
            observedAt:
              type: string
              format: date-time
            receivedAt:
              type: string
              format: date-time
            staleAfterSeconds:
              type: integer
              nullable: true
            isStale:
              type: boolean